Refactored and Optimized the TouchInputManager and UIRaycast
UIRaycast will now return List<RaycastResult> instead of IInputListener. TouchInputManager keeps a cache of the subscriber's gameObject and IInputListener component. We only consider the raycast result at the very front and ignore every other behind it.
IInputListener subscribe to GameActions in the Activate and DeActivate functions instead of doing it themselves via OnEnable. The InputControlHandler has a list of all IInputListeners in the scene and controls which ones become active based on their listenerGroup property (Which is a string)